Refactor utils packages
