+++ internal/release/promote/perform.go
@@ -3,9 +3,9 @@ package promote
 import (
 	"fmt"
 	"github.com/TwiN/go-color"
-	"github.com/nestoca/joy-cli/internal/colors"
 	"github.com/nestoca/joy-cli/internal/git"
 	"github.com/nestoca/joy-cli/internal/release"
+	"github.com/nestoca/joy-cli/internal/utils/colors"
 	"github.com/nestoca/joy-cli/internal/yml"
 	"strings"
 )
