+++ internal/utils/home/resolve_path.go
@@ -1,4 +1,4 @@
-package utils
+package home
 
 import (
 	"fmt"
